Add AppImage packaging for Linux (#13)
* Fix filename for libweaver.so

CMake was adding an extra "lib" prefix on linux, with filename resulting
in "liblibweaver.so". Now it generates it as "libweaver.so."

* Add Linux workflow

The workflow runs on ubuntu and creates 2 arficacts: libweaver.so and an
AppImage containing the SI Editor GUI app.

Also adding a .desktop file as required for building an AppImage.